Removed manual Zephyr Badge, set scraper to also pick up "Johto", and allowed fractional offsets

+++ src/tpp-progress.ts
@@ -266,7 +266,7 @@ function applyScale(ppd?: number) {
         chart.style.backgroundImage = 'url("' + makeGrid(ppd) + '")';
     });
     $find(fakeQuery(".progressChart .ruler"), ".stop").forEach(ruler=> ruler.forEach((stop, i) => {
-        var offset = parseInt($(stop).parents('.progressChart').data('offset') || '0');
+        var offset = parseFloat($(stop).parents('.progressChart').data('offset') || '0');
         stop.style.left = (i + offset) * ppd + "px";
     }));
     fakeQuery(".progressChart > .run").forEach(run=> {
@@ -291,7 +291,7 @@ function applyScale(ppd?: number) {
         if (settings["explode"]) {
             staggerStackedEvents(events.filter(e=> e.style.display != "none"), run.offsetHeight);
         }
-        var offset = parseInt($(run).parents('.progressChart').data('offset') || '0');
+        var offset = parseFloat($(run).parents('.progressChart').data('offset') || '0');
         run.style.marginLeft = offset * ppd + "px";
         $(run).find('.hosts').first().css('margin-left', -offset * ppd + "px");
     });
